central-server/backend/requirements.txt
Flask
psycopg2-binary 